split tests, use _ServiceWatcher directly
